Bullet, Zoom, Shake, Gui
Maddy did part of the GUI on my laptop and under my name because Matt wasn't there to let us in to use the computers that are in the room. Todd helped me with the camera zoom and the player shake. Todd also did a backup bullet thing so that Helen and I can start coding the firing aspect of the player. Got a lot done, was very productive.