Camera Controller and Main Menu Screen
Took the code that controlled the camera zooming and put it into it's own script to make sure that if we got rid of the player then that part would still be able to run. Also made a main menu screen to start the game up with and I also made it so if you click the play game button then it will take you to the right scene to play the game in.